requirements.txt
flask
mongoengine
python-dotenv
flask-jwt-extended